Improve Pick Team
Some searches seem to not provide good matches... Might be that the
search is based on the team description, and not on the team name?

This commit adds a lookup team by id, in case the name search doesn't
give any matches.
Also sort list of teams, based on search term.

- Add Search by Team Id
- Sort team name matches