Add Pick Team by Arena id
Previous team and battle selections relied on first picking a team,
and then checking which team battles the team are participating in.
For Lichess to list team battle participation by team, a team leader
must have joined the team battle.
So (in the rare case) when a team leader isn't participating, it wasn't
possible to find the arena.
If the user already knows an arena id, it is possible to list which
teams are participating in the arena (also teams without participating
team leader).
This commit adds the possibility to pick a team (and arena) based on
arena id.