Avoid member game results outside of team battle
Typically game results will be from games in the team battle.
But players could play games outside of the team battle!
If a member withdraws from the team battle, games of that member won't
be monitored - but if the member never plays a game, their withdraw
status won't be set - i.e it can appear they are waiting to get paired.
Since they can get paired, they're added to the game result monitoring,
and if they start games outside of the team battle, their game result
would be counted as being a result for the team battle.

This commit adds additional checks of some available indicators to avoid
reporting results from outside the team battle:
- Game rated/casual is same as team battle
- Game time control is same as team battle
- Game participants have both joined the team battle

Note, to be certain the game result is from the monitored team battle,
one could make an additional API request with the game id,
but above checks might be good enough to avoid that.

    boolean fromTeamBattle = client.games().byGameId(id)
        .map(Game::tournament)
        .map(tour -> tour instanceof Some(String tourId)
                     && teamBattleId.equals(tourId))
        .orElse(false);